apps/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-environ==0.12.0
gunicorn==23.0.0
packaging==25.0
psycopg==3.2.11
psycopg-binary==3.2.11
sqlparse==0.5.3
typing_extensions==4.15.0
